requirements.txt
aiogram==2.25.1
python-dotenv==1.0.1
openpyxl==3.1.2
pandas==2.1.4
et-xmlfile==1.1.0 
Pillow==10.3.0